open tensorrt unittest
