+++ test/tensorrt/CMakeLists.txt
@@ -1,10 +1,11 @@
-# Open after the unilateral issue is resolved
-# file(
-#   GLOB TEST_OPS
-#   RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}"
-#   "test_*.py")
-# string(REPLACE ".py" "" TEST_OPS "${TEST_OPS}")
+if(NOT WIN32 AND TENSORRT_FOUND)
+  file(
+    GLOB TEST_OPS
+    RELATIVE "${CMAKE_CURRENT_SOURCE_DIR}"
+    "test_*.py")
+  string(REPLACE ".py" "" TEST_OPS "${TEST_OPS}")
 
-# foreach(TEST_OP ${TEST_OPS})
-#   py_test_modules(${TEST_OP} MODULES ${TEST_OP})
-# endforeach()
+  foreach(TEST_OP ${TEST_OPS})
+    py_test_modules(${TEST_OP} MODULES ${TEST_OP})
+  endforeach()
+endif()
